+++ src/modules/helpers/builder/plugins/npm-modules.ts
@@ -1,9 +1,9 @@
 import type { OnResolveArgs, Plugin } from 'esbuild'
 
 /**
- * esbuild plugin to ignore resolution errors for default modules.
+ * esbuild plugin to ignore resolution errors for npm modules.
  *
- * This plugin intercepts the module resolution process and marks certain modules
+ * This plugin intercepts the module resolution process and marks npm modules
  * (like `esbuild` and `javascript-obfuscator`) as external, meaning they will not
  * be bundled and their resolution will be ignored during the build process.
  * This helps to avoid errors related to modules that cannot be resolved or are
@@ -11,13 +11,14 @@ import type { OnResolveArgs, Plugin } from 'esbuild'
  *
  * @returns {Plugin} esbuild plugin to ignore resolution errors for specific modules.
  */
-export function ignoreDefaultModulesPlugin(): Plugin {
+export function npmModulesPlugin(libraries: string[]): Plugin {
+  const externals = [...defaultNpmModules, ...libraries]
   return {
-    name: 'ignore-default-modules',
+    name: 'npm-modules',
     setup(build) {
       build.onResolve({ filter: /.*/ }, (args: OnResolveArgs) => {
         // Ignore resolution for specific problematic modules
-        if (ignoredDefaultModules.includes(args.path)) {
+        if (externals.includes(args.path)) {
           return { external: true, path: `npm:${args.path}` } // Send as external npm
         }
         return null
@@ -26,4 +27,4 @@ export function ignoreDefaultModulesPlugin(): Plugin {
   }
 }
 
-export const ignoredDefaultModules = ['esbuild', 'javascript-obfuscator']
+export const defaultNpmModules = ['esbuild', 'javascript-obfuscator']