+++ src/modules/helpers/zanix/projects/library.ts
@@ -1,12 +1,13 @@
 import type { ZanixLibrarySrcTree } from 'typings/zanix.ts'
 
 import { ZanixTree } from 'modules/helpers/zanix/base-tree.ts'
+import { MAIN_MODULE } from 'utils/constants.ts'
 import { join } from '@std/path'
 
 export const getLibrarySrcTree = (root: string): ZanixLibrarySrcTree => {
-  const mainRoot = join(root, 'src/modules')
+  const startingPoint = join(root, 'src/modules')
 
-  return ZanixTree.create<ZanixLibrarySrcTree>(mainRoot, {
-    templates: { base: { files: ['mod.ts'], jsr: '@zanix/utils' } },
+  return ZanixTree.create<ZanixLibrarySrcTree>({ startingPoint, baseRoot: root }, {
+    templates: { base: { files: [MAIN_MODULE], jsr: '@zanix/utils' } },
   })
 }